+++ src/enums/PayableEntity.ts
@@ -1,7 +1,8 @@
 enum PayableEntity {
   ApplePay = 'applebtn',
   GooglePay = 'googlebtn',
-  Form = 'form'
+  Form = 'form',
+  Resign = 'resign'
 }
 
 export default PayableEntity