+++ src/enums/ResignFieldName.ts
@@ -1,0 +1,5 @@
+enum ResignFieldName {
+  ResignCvv = 'resignCvv'
+}
+
+export default ResignFieldName