+++ src/interfaces/messages/InteractionMessage.ts
@@ -5,24 +5,41 @@ import AdditionalFieldName from '../../enums/AdditionalFieldName'
 import InteractionType from '../../enums/InteractionType'
 import InteractionTargetType from '../../enums/InteractionTargetType'
 import MessageType from '../../enums/MessageType'
+import ResignFieldName from '../../enums/ResignFieldName'
 
 interface ControlState {
   isValid: boolean
   isTouched: boolean
 }
 
-export default interface InteractionMessage
-  extends Message<MessageType.Interaction> {
+type InteractionMessageFormState =
+  | {
+      cardForm: {
+        fields: Record<
+          FieldName.CardCvv | FieldName.CardNumber | FieldName.CardExpiryDate,
+          ControlState
+        > &
+          Partial<Record<AdditionalFieldName | 'cardHolder', ControlState>>
+      } & ControlState
+    }
+  | {
+      resignForm: {
+        fields: Record<ResignFieldName, ControlState>
+      } & ControlState
+    }
+
+type InteractionMessage = Message<MessageType.Interaction> & {
   target: {
     type: InteractionTargetType
-    name: FieldName | AdditionalFieldName | 'submit' | 'googlePay' | 'applePay'
+    name:
+      | FieldName
+      | ResignFieldName
+      | AdditionalFieldName
+      | 'submit'
+      | 'googlePay'
+      | 'applePay'
     interaction: InteractionType
   }
-  cardForm: {
-    fields: Record<
-      FieldName.CardCvv | FieldName.CardNumber | FieldName.CardExpiryDate,
-      ControlState
-    > &
-      Partial<Record<AdditionalFieldName | 'cardHolder', ControlState>>
-  } & ControlState
-}
+} & InteractionMessageFormState
+
+export default InteractionMessage