+++ src/interfaces/messages/MountedMessage.ts
@@ -3,5 +3,5 @@ import Message from './Message'
 import MessageType from '../../enums/MessageType'
 
 export default interface MountedMessage extends Message<MessageType.Mounted> {
-  entity: 'applebtn' | 'googlebtn' | 'form'
+  entity: 'applebtn' | 'googlebtn' | 'form' | 'resign'
 }